--- java-tools/OsmAndServer/src/main/java/net/osmand/server/api/searchtest/dto/Function.java
@@ -1,4 +1,0 @@
-package net.osmand.server.api.searchtest.dto;
-
-public record Function(String name, String description, Param[] param) {
-}
