--- java-tools/OsmAndServer/src/main/java/net/osmand/server/api/searchtest/dto/Param.java
@@ -1,4 +1,0 @@
-package net.osmand.server.api.searchtest.dto;
-
-public record Param(String name, String type, String description) {
-}
