--- java-tools/OsmAndServer/src/main/java/net/osmand/server/api/searchtest/dto/TestCaseStatus.java
@@ -1,12 +1,0 @@
-package net.osmand.server.api.searchtest.dto;
-
-import net.osmand.server.api.searchtest.entity.TestCase;
-
-public record TestCaseStatus(
-		TestCase.Status status,
-		long processed,
-		long failed,
-		long filtered,
-		long empty,
-		long duration) {
-}
