--- java-tools/OsmAndServer/src/main/java/net/osmand/server/api/searchtest/entity/GenResult.java
@@ -1,9 +1,0 @@
-package net.osmand.server.api.searchtest.entity;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.Table;
-
-@Entity
-@Table(name = "gen_result")
-public class GenResult extends Result {
-}
